Fixes label instruction at end and generates assembly
When a label instruction is at the end of a fragment,
instr2graph creates a edge to a node that doesn't exist,
creating a index out of bound exception when calculating
liveness. To fix this, an empty node is created if a
label is at the end of a fragment.

Adds runtime.s and fixes MIPS formatting so that SPIM
doesn't complain. Strings and instructions are separated
into .data and .text sections now.